Moved token range support into separate file. Stop iteration early in findInstancesForKey, if possible.
