+++ ring/ring.go
@@ -9,7 +9,6 @@ import (
 	"math"
 	"math/rand"
 	"net/http"
-	goslices "slices"
 	"sync"
 	"time"
 
@@ -364,7 +363,9 @@ func (r *Ring) Get(key uint32, op Operation, bufDescs []InstanceDesc, bufHosts, 
 		return ReplicationSet{}, ErrEmptyRing
 	}
 
-	instances, err := r.findInstancesForKey(key, op, bufDescs, bufHosts, bufZones, nil)
+	instances, err := r.findInstancesForKey(key, op, bufDescs, bufHosts, bufZones, func(d InstanceDesc) (include, keepGoing bool) {
+		return true, true
+	})
 	if err != nil {
 		return ReplicationSet{}, err
 	}
@@ -381,8 +382,8 @@ func (r *Ring) Get(key uint32, op Operation, bufDescs []InstanceDesc, bufHosts, 
 }
 
 // Returns instances for given key and operation. Instances are not filtered through ReplicationStrategy.
-// InstanceFilter can ignore uninteresting instances that would otherwise be part of the output.
-func (r *Ring) findInstancesForKey(key uint32, op Operation, bufDescs []InstanceDesc, bufHosts []string, bufZones []string, instanceFilter func(d InstanceDesc) bool) ([]InstanceDesc, error) {
+// InstanceFilter can ignore uninteresting instances that would otherwise be part of the output, and can also stop search early.
+func (r *Ring) findInstancesForKey(key uint32, op Operation, bufDescs []InstanceDesc, bufHosts []string, bufZones []string, instanceFilter func(d InstanceDesc) (include, keepGoing bool)) ([]InstanceDesc, error) {
 	var (
 		n            = r.cfg.ReplicationFactor
 		instances    = bufDescs[:0]
@@ -433,8 +434,15 @@ func (r *Ring) findInstancesForKey(key uint32, op Operation, bufDescs []Instance
 			distinctZones = append(distinctZones, info.Zone)
 		}
 
-		if instanceFilter == nil || instanceFilter(instance) {
+		include, keepGoing := true, true
+		if instanceFilter != nil {
+			include, keepGoing = instanceFilter(instance)
+		}
+		if include {
 			instances = append(instances, instance)
+		}
+		if !keepGoing {
+			break
 		}
 	}
 	return instances, nil
@@ -1095,103 +1103,6 @@ func (op Operation) ShouldExtendReplicaSetOnState(s InstanceState) bool {
 // All states are healthy, no states extend replica set.
 var allStatesRingOperation = Operation(0x0000ffff)
 
-func (r *Ring) GetTokenRangesForInstance(instanceID string) ([]uint32, error) {
-	r.mtx.RLock()
-	defer r.mtx.RUnlock()
-
-	instance, ok := r.ringDesc.Ingesters[instanceID]
-	if !ok {
-		return nil, ErrInstanceNotFound
-	}
-	if instance.Zone == "" {
-		return nil, errors.New("zone not set")
-	}
-
-	rf := r.cfg.ReplicationFactor
-	numZones := len(r.ringTokensByZone)
-
-	// To simplify computation of token ranges, we currently only support case where zone-awareness is enabled,
-	// and replicaction factor is equal to number of zones.
-	if !r.cfg.ZoneAwarenessEnabled || rf != numZones {
-		// if zoneAwareness is disabled we treat the whole ring as one big zone, and would
-		// need to walk the ring backwards looking for RF-1 tokens from other instances to determine the range
-		// ignore this for now
-		return nil, errors.New("can't use ring configuration for computing token ranges")
-	}
-
-	// at this point zone-aware replication is enabled, and rf == numZones
-	// this means that we will write to one replica in each zone, so we can just consider the zonal ring for our instance
-	subringTokens, ok := r.ringTokensByZone[instance.Zone]
-	if !ok || len(subringTokens) == 0 {
-		return nil, errors.New("no tokens for zone")
-	}
-
-	ranges := make([]uint32, 0, 2*(len(instance.Tokens)+1)) // 1 range (2 values) per token + one additional if we need to split the rollover range
-	var rangeEnd uint32
-
-	// if this instance claimed the first token, it owns the wrap-around range, which we'll break into two separate ranges
-	firstToken := subringTokens[0]
-	firstTokeninfo, ok := r.ringInstanceByToken[firstToken]
-	if !ok {
-		// This should never happen unless there's a bug in the ring code.
-		return nil, ErrInconsistentTokensInfo
-	}
-
-	if firstTokeninfo.InstanceID == instanceID {
-		// we'll start by looking for the beginning of the range that ends with math.MaxUint32
-		rangeEnd = math.MaxUint32
-	}
-
-	// walk the ring backwards, alternating looking for ends and starts of ranges
-	for i := len(subringTokens) - 1; i > 0; i-- {
-		token := subringTokens[i]
-		info, ok := r.ringInstanceByToken[token]
-		if !ok {
-			// This should never happen unless a bug in the ring code.
-			return nil, ErrInconsistentTokensInfo
-		}
-
-		if rangeEnd == 0 {
-			// we're looking for the end of the next range
-			if info.InstanceID == instanceID {
-				rangeEnd = token - 1
-			}
-		} else {
-			// we have a range end, and are looking for the start of the range
-			if info.InstanceID != instanceID {
-				ranges = append(ranges, rangeEnd, token)
-				rangeEnd = 0
-			}
-		}
-	}
-
-	// finally look at the first token again
-	// - if we have a range end, check if we claimed token 0
-	//   - if we don't, we have our start
-	//   - if we do, the start is 0
-	// - if we don't have a range end, check if we claimed token 0
-	//   - if we don't, do nothing
-	//   - if we do, add the range of [0, token-1]
-	//     - BUT, if the token itself is 0, do nothing, because we don't own the tokens themselves (we should be covered by the already added range that ends with MaxUint32)
-
-	if rangeEnd == 0 {
-		if firstTokeninfo.InstanceID == instanceID && firstToken != 0 {
-			ranges = append(ranges, firstToken-1, 0)
-		}
-	} else {
-		if firstTokeninfo.InstanceID == instanceID {
-			ranges = append(ranges, rangeEnd, 0)
-		} else {
-			ranges = append(ranges, rangeEnd, firstToken)
-		}
-	}
-
-	// Ensure returned ranges are sorted.
-	goslices.Sort(ranges)
-
-	return ranges, nil
-}
-
 func (r *Ring) NumberOfKeysOwnedByInstance(keys []uint32, op Operation, instanceID string) (int, error) {
 	r.mtx.RLock()
 	defer r.mtx.RUnlock()
@@ -1206,8 +1117,12 @@ func (r *Ring) NumberOfKeysOwnedByInstance(keys []uint32, op Operation, instance
 
 	result := 0
 	for _, tok := range keys {
-		i, err := r.findInstancesForKey(tok, op, bufDescs, bufHosts, bufZones, func(d InstanceDesc) bool {
-			return d.Id == instanceID
+		i, err := r.findInstancesForKey(tok, op, bufDescs, bufHosts, bufZones, func(d InstanceDesc) (include, keepGoing bool) {
+			if d.Id == instanceID {
+				// If we've found our instance, we can stop.
+				return true, false
+			}
+			return false, true
 		})
 		if err != nil {
 			return 0, err
